+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 
 ## Unpublished
+### Changed
+- group/list switch selection in attendance record is now cached in the session
+### Fixed
+- attendance record shows correct member sets after filter change (without the need of a reload)
 
 ## v0.7.5 - 2024-04-06
 ### Added